splits submit page into a few different files
